Build configuration - Changed some ints to NSIntegers. - Coverage builds should now work.
git-svn-id: http://basetenframework.org/svn/trunk@1336 fe1aa653-7022-0410-af67-bafd91c7e665

+++ BaseTenAssistant/Sources/MKCAlternativeDataCellColumn.m
@@ -38,7 +38,7 @@
 
 @implementation MKCAlternativeDataCellColumn
 
-- (id) dataCellForRow: (int) rowIndex
+- (id) dataCellForRow: (NSInteger) rowIndex
 {
     id retval = nil;
     if (-1 == rowIndex)
